Describe as "offset in document"
The position returned is relative to the document, not the viewport. The
viewport (see http://www.quirksmode.org/mobile/viewports.html#link9) is
the area visible to the user (i.e., the browser window), so it would be
expected for the returned value to change as the user scrolls the
document. This is *not* the case for the value returned by this module.